requirements.txt
asgiref==3.8.1
backports.zoneinfo==0.2.1
beautifulsoup4==4.12.3
certifi==2024.6.2
charset-normalizer==3.3.2
crispy-forms-materialize==0.2
Django==4.2.13
django-crispy-forms==2.1
django-health-check==3.18.2
django-materializecss-form==1.1.17
django-pwa==1.1.0
djangorestframework==3.15.1
google==3.0.0
idna==3.7
psutil==5.9.8
requests==2.32.3
soupsieve==2.5
sqlparse==0.5.0
typing-extensions==4.12.2
urllib3==2.2.1
